+++ savanna/src/com/savanna/model/command/member/MemDetailCommand.java
@@ -1,10 +1,6 @@
 package com.savanna.model.command.member;
 
 import java.io.IOException;
-import java.io.PrintWriter;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
@@ -13,7 +9,6 @@ import javax.servlet.http.HttpServletResponse;
 import com.savanna.model.command.Command;
 import com.savanna.model.dao.MemberDAO;
 import com.savanna.model.vo.MemberVO;
-import com.savanna.model.vo.Paging;
 
 public class MemDetailCommand implements Command{
 	String id;
@@ -28,14 +23,9 @@ public class MemDetailCommand implements Command{
 			throws ServletException, IOException {
 		request.setCharacterEncoding("UTF-8");
 		
-
-		
 		MemberVO vo = MemberDAO.memDetail(id);
 //		System.out.println(vo);
-
 		request.setAttribute("vo", vo);
-
-		
 		
 		return "member/memDetail.jsp";
 	}
