+++ savanna/src/com/savanna/model/command/member/MemListCommand.java
@@ -1,10 +1,6 @@
 package com.savanna.model.command.member;
 
 import java.io.IOException;
-import java.io.PrintWriter;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
@@ -12,11 +8,8 @@ import javax.servlet.http.HttpServletResponse;
 
 import com.savanna.model.command.Command;
 import com.savanna.model.command.paging.ListPaging;
-import com.savanna.model.dao.BookDAO;
 import com.savanna.model.dao.MemberDAO;
-import com.savanna.model.vo.BookVO;
 import com.savanna.model.vo.MemberVO;
-import com.savanna.model.vo.Paging;
 
 public class MemListCommand implements Command{
 	String cPage;
@@ -31,7 +24,7 @@ public class MemListCommand implements Command{
 		request.setCharacterEncoding("UTF-8");
 		response.setContentType("text/html;charset=UTF-8");
 		
-		new ListPaging<BookVO>().PagingDistributor(new BookDAO(), request);
+		new ListPaging<MemberVO>().PagingDistributor(new MemberDAO(), request);
 		request.setAttribute("cPage", cPage);
 
 		return "member/memList.jsp";
