--- savanna/src/com/savanna/model/vo/Paging.java
@@ -1,101 +1,0 @@
-package com.savanna.model.vo;
-
-//회원목록의 페이징 처리를 위한 값을 저장 관리
-public class Paging {
-	private int nowPage = 1; //현재 페이지
-	private int nowBlock = 1; //현재 블록(페이지 담는 단위)
-	
-	private int numPerPage = 10; //하나의 페이지에 표시할 회원 수
-	private int pagePerBlock = 5; //블록당 표시하는 페이지 개수
-	
-	private int totalMember = 0; //총 회원수
-	private int totalPage = 0; //전체 페이지 개수
-	private int totalBlock = 0; //전체 블록 개수
-	
-	private int begin = 0; //현재 페이지상의 시작 회원번호
-	private int end = 0; //현재 페이지상의 마지막 회원번호
-	
-	private int beginPage = 0; //현재 블록의 시작 페이지 번호
-	private int endPage = 0; //현재 블록의 끝 페이지 번호
-	
-	//전체 페이지 개수 구하기
-	//totalMember 값을 페이지당 표시할 회원의 개수 값을 나누고,
-	//나머지가 있으면 페이지 하나 더 추가
-	public void setTotalPage() {
-		totalPage = totalMember / numPerPage;
-		if (totalMember % numPerPage > 0) totalPage++;
-	}
-	
-	public int getNowPage() {
-		return nowPage;
-	}
-	public void setNowPage(int nowPage) {
-		this.nowPage = nowPage;
-	}
-	public int getNowBlock() {
-		return nowBlock;
-	}
-	public void setNowBlock(int nowBlock) {
-		this.nowBlock = nowBlock;
-	}
-	public int getNumPerPage() {
-		return numPerPage;
-	}
-	public void setNumPerPage(int numPerPage) {
-		this.numPerPage = numPerPage;
-	}
-	public int getPagePerBlock() {
-		return pagePerBlock;
-	}
-	public void setPagePerBlock(int pagePerBlock) {
-		this.pagePerBlock = pagePerBlock;
-	}
-	public int getTotalMember() {
-		return totalMember;
-	}
-	public void setTotalMember(int totalMember) {
-		this.totalMember = totalMember;
-	}
-	public int getTotalPage() {
-		return totalPage;
-	}
-	public void setTotalPage(int totalPage) {
-		this.totalPage = totalPage;
-	}
-	public int getTotalBlock() {
-		return totalBlock;
-	}
-	public void setTotalBlock(int totalBlock) {
-		this.totalBlock = totalBlock;
-	}
-	public int getBegin() {
-		return begin;
-	}
-	public void setBegin(int begin) {
-		this.begin = begin;
-	}
-	public int getEnd() {
-		return end;
-	}
-	public void setEnd(int end) {
-		this.end = end;
-	}
-
-	public int getBeginPage() {
-		return beginPage;
-	}
-
-	public void setBeginPage(int beginPage) {
-		this.beginPage = beginPage;
-	}
-
-	public int getEndPage() {
-		return endPage;
-	}
-
-	public void setEndPage(int endPage) {
-		this.endPage = endPage;
-	}
-
-
-}
